создание папки controllers и перемещение файлов
