+++ controllers/contacts.js
@@ -1,22 +1,15 @@
-const express = require("express");
-const router = express.Router();
-const Contacts = require("../../model/contacts");
-const {
-  validateAddContact,
-  validateUpdateContact,
-  validateStatusFavoriteContact,
-} = require("./validation");
+const Contacts = require("../model/contacts");
 
-router.get("/", async (req, res, next) => {
+const getAll = async (req, res, next) => {
   try {
     const contacts = await Contacts.listContacts();
     return res.json({ status: "success", code: 200, data: { contacts } });
   } catch (error) {
     next(error);
   }
-});
+};
 
-router.get("/:contactId", async (req, res, next) => {
+const getById = async (req, res, next) => {
   try {
     const contact = await Contacts.getContactById(req.params.contactId);
     if (contact) {
@@ -30,9 +23,9 @@ router.get("/:contactId", async (req, res, next) => {
   } catch (error) {
     next(error);
   }
-});
+};
 
-router.post("/", validateAddContact, async (req, res, next) => {
+const create = async (req, res, next) => {
   if (JSON.stringify(req.body) === "{}") {
     return res.status(200).json({
       status: "no body",
@@ -48,9 +41,9 @@ router.post("/", validateAddContact, async (req, res, next) => {
   } catch (error) {
     next(error);
   }
-});
+};
 
-router.delete("/:contactId", async (req, res, next) => {
+const remove = async (req, res, next) => {
   try {
     const contact = await Contacts.removeContact(req.params.contactId);
     if (contact) {
@@ -64,9 +57,9 @@ router.delete("/:contactId", async (req, res, next) => {
   } catch (error) {
     next(error);
   }
-});
+};
 
-router.put("/:contactId", validateUpdateContact, async (req, res, next) => {
+const update = async (req, res, next) => {
   if (JSON.stringify(req.body) === "{}") {
     return res
       .status(200)
@@ -88,37 +81,37 @@ router.put("/:contactId", validateUpdateContact, async (req, res, next) => {
   } catch (error) {
     next(error);
   }
-});
+};
 
-router.patch(
-  "/:contactId/favorite",
-  validateStatusFavoriteContact,
-  validateUpdateContact,
-  async (req, res, next) => {
-    if (JSON.stringify(req.body) === "{}") {
-      return res.status(200).json({
-        status: "no body",
-        code: 400,
-        message: "missing field favorite",
-      });
-    }
-    try {
-      const contact = await Contacts.updateContact(
-        req.params.contactId,
-        req.body
-      );
-      if (contact) {
-        return res
-          .status(200)
-          .json({ status: "success", code: 200, data: { contact } });
-      }
-      return res
-        .status(404)
-        .json({ status: "error", code: 404, message: "Not Found" });
-    } catch (error) {
-      next(error);
-    }
-  }
-);
-
-module.exports = router;
+// async (req, res, next) => {
+//   if (JSON.stringify(req.body) === "{}") {
+//     return res.status(200).json({
+//       status: "no body",
+//       code: 400,
+//       message: "missing field favorite",
+//     });
+//   }
+//   try {
+//     const contact = await Contacts.updateContact(
+//       req.params.contactId,
+//       req.body
+//     );
+//     if (contact) {
+//       return res
+//         .status(200)
+//         .json({ status: "success", code: 200, data: { contact } });
+//     }
+//     return res
+//       .status(404)
+//       .json({ status: "error", code: 404, message: "Not Found" });
+//   } catch (error) {
+//     next(error);
+//   }
+// };
+module.exports = {
+  getAll,
+  getById,
+  create,
+  remove,
+  update,
+};
